Upload openklem cleaning code

+++ openklem-algo/src/main/java/com/geomaticaeambiente/openjump/klem/parallel/ExecutorBuilder.java
@@ -13,7 +13,7 @@ public class ExecutorBuilder {
         
         int threadCount = 1; //Runtime.getRuntime().availableProcessors();
         executorService = Executors.newFixedThreadPool(threadCount);
-        rowPerStripeCount = (int) rowCount / threadCount;
+        rowPerStripeCount = rowCount / threadCount;
         if(rowPerStripeCount == 0) {
             rowPerStripeCount = 1;
         }
